replaced and renamed api.ts to utils/usertils.ts
